Defining variable body on line 85
According to the method logic, if an empty DataFrame is passed as an argument, the while loop will never run and thus line 141 will throw an error "Variable referenced before assignment" as variable body is referenced which is defined in the loop that has not been run. Defining variable body before the loop will remove this error.